Made Spawning Truly Consistent + Enemy-Crate Treaty. Also Buffed The Darn Grenade Launcher!!
Spawning is now truly dependent on solely the amount of time that has
passed. Enemies now agree not to harm crates unless the crate is
directly in their way for a prolonged time, so collateral damage should
be impossible. Also, I buffed the darn grenade launcher!!... it's still
a "challenge weapon" nonetheless.